@@ -12,38 +12,32 @@ dependencies:
     sdk: flutter
 
   ## BLOC
+  bloc: ^0.14.4
+  flutter_bloc: ^0.19.0
   rxdart: ^0.22.0
-  bloc: ^0.14.2
-  flutter_bloc: ^0.16.0
 
   ## FIREBASE
-#  firebase_core: ^0.4.0+1
-  cloud_firestore: ^0.12.5
-  cloud_functions: ^0.4.0
-  firebase_auth: ^0.11.1+6
+  cloud_firestore: ^0.12.7
+  cloud_functions: ^0.4.0+2
+  firebase_analytics: ^3.0.3
+  firebase_auth: ^0.11.1+7
   firebase_remote_config: ^0.2.0+3
-  firebase_analytics: ^3.0.2
-  google_sign_in: ^4.0.1+3
+  # google_sign_in: ^4.0.1+3
 
   ## CONFIGURATION
+  connectivity: ^0.4.3+4
   package_info: ^0.4.0+4
-  connectivity: ^0.4.2+1
-
-  ## LOGGING
-
+  # device_info
 
   ## UTILS
-  dio: ^2.1.3
-  flare_flutter: ^1.3.6
-#  logging: ^0.11.3+2
-  logger: ^0.6.0
-  intl: ^0.15.8
+  ansicolor: ^1.0.2
   date_utils: ^0.1.0+2
-  uuid: ^2.0.1
+  dio: ^2.1.13
+  flare_flutter: ^1.5.4
+  intl: ^0.15.8
   merge_map: ^1.0.2
   shimmer: ^1.0.0
-  ansicolor: ^1.0.2
-#  flutter_svg: ^0.13.0+2 FIXME: only supports SDK version <2.0.0
+  uuid: ^2.0.2
 
 dev_dependencies:
   flutter_test:
@@ -59,10 +53,10 @@ dev_dependencies:
 
   ## BUILT VALUE
   build: ^1.1.4
-  build_runner: ^1.5.0
-  built_value: ^6.6.0
+  build_runner: ^1.6.1
   built_collection: ^4.2.2
-  built_value_generator: ^6.6.0
+  built_value: ^6.7.0
+  built_value_generator: ^6.7.0
 
 #  pedantic: ^1.7.0
   flutter_launcher_icons: ^0.7.2
