@@ -9,11 +9,21 @@ class LoggingBloc extends Bloc<LoggingEvent, LoggingState> {
   @override
   LoggingState get initialState => LoggingState();
 
+  // Custom print to avoid mobile phone console truncation
+  // https://github.com/flutter/flutter/issues/22665#issuecomment-458186456
+  void printWrapped(Object object, {int chunkSize = 800}) {
+    final line = "$object";
+    final pattern = RegExp(".{1,$chunkSize}");
+    for (var match in pattern.allMatches(line)) {
+      print(match.group(0));
+    }
+  }
 
   @override
   void onEvent(LoggingEvent event) {
 
   }
+
 
 
   @override
@@ -23,6 +33,9 @@ class LoggingBloc extends Bloc<LoggingEvent, LoggingState> {
 
   @override
   void onError(Object error, StackTrace stacktrace) {
+    print("ERROR OCCURED IN LOGGING - $error");
+    print(stacktrace);
+
     // FIXME: this only catches errors from within mapEventToState, would this catch the constructor's dispatches?
     // strait up throw errors into logging -- better to throw error as own type of log!
   }
@@ -31,47 +44,49 @@ class LoggingBloc extends Bloc<LoggingEvent, LoggingState> {
   Stream<LoggingState> mapEventToState(LoggingEvent event) async* {
     final loggingBuilder = currentState.toBuilder();
 
-    if (event is SendAnonymousLogsToDeveloper) {
+    if (event is SendLogsToDeveloper) {
       print("SENT ${currentState.logs.length} LOGS TO DEVELOPER"); // TODO
-      return;
+//      return;
     }
 
     if (event is LogMessage) {
       final messageLog = MessageLog((b) => b
-        ..level = event.level
-        ..text = event.text
+        ..message = event.message
         ..datetime = DateTime.now()
+        ..level = event.level
       );
 
       loggingBuilder.logs.add(messageLog);
-
-      final pen = AnsiPen()..white(bold: true)..rgb(r: 1.0, g: 0.8, b: 0.2);
-      print(pen("Bright white foreground") + " this text is default fg/bg");
-
-      print("LOGGING MESSAGE ---");
-      print(messageLog);
+      printWrapped(messageLog);
+//      final pen = AnsiPen()..white(bold: true)..rgb(r: 1.0, g: 0.8, b: 0.2);
+//      print(pen("Bright white foreground") + " this text is default fg/bg");
     }
 
     if (event is LogError) {
       final errorLog = ErrorLog((b) => b
+        ..message = event.message
+        ..datetime = DateTime.now()
         ..level = event.level
         ..error = event.error
         ..stacktrace = event.stacktrace
-        ..datetime = DateTime.now()
       );
 
       loggingBuilder.logs.add(errorLog);
+      printWrapped(errorLog);
     }
 
     if (event is LogBlocTransition) {
+
       final blocTransition = BlocTransitionLog((b) => b
+        ..message = event.message
+        ..datetime = DateTime.now()
         ..currentState = event.currentState
         ..event = event.event
         ..nextState = event.nextState
-        ..datetime = DateTime.now()
       );
 
       loggingBuilder.logs.add(blocTransition);
+      printWrapped(blocTransition);
 
       //    logger.i("${transition.event}");
 //    logger.i("${transition.event.runtimeType}");
@@ -88,60 +103,62 @@ class LoggingBloc extends Bloc<LoggingEvent, LoggingState> {
   /// Instead of `LoggingBloc().dispatch(...)` for every log;
 
   ///
-  void expectedError(Object error, StackTrace stacktrace) {
+  void expectedError(String message, Object error, StackTrace stacktrace) {
     dispatch(LogError((b) => b
       ..level = ErrorLoggingLevel.expected
+      ..message = message
       ..error = error
       ..stacktrace = stacktrace
     ));
   }
 
   ///
-  void unexpectedError(Object error, StackTrace stacktrace) {
+  void unexpectedError(String message, Object error, StackTrace stacktrace) {
     dispatch(LogError((b) => b
       ..level = ErrorLoggingLevel.unexpected
+      ..message = message
       ..error = error
       ..stacktrace = stacktrace
     ));
   }
 
   ///
-  void verbose(String text) {
+  void verbose(String message) {
     dispatch(LogMessage((b) => b
       ..level = MessageLoggingLevel.verbose
-      ..text = text
+      ..message = message
     ));
   }
 
   ///
-  void debug(String text) {
+  void debug(String message) {
     dispatch(LogMessage((b) => b
       ..level = MessageLoggingLevel.debug
-      ..text = text
+      ..message = message
     ));
   }
 
   ///
-  void info(String text) {
+  void info(String message) {
     dispatch(LogMessage((b) => b
       ..level = MessageLoggingLevel.info
-      ..text = text
+      ..message = message
     ));
   }
 
   ///
-  void warning(String text) {
+  void warning(String message) {
     dispatch(LogMessage((b) => b
       ..level = MessageLoggingLevel.warning
-      ..text = text
+      ..message = message
     ));
   }
 
   ///
-  void error(String text) {
+  void error(String message) {
     dispatch(LogMessage((b) => b
       ..level = MessageLoggingLevel.error
-      ..text = text
+      ..message = message
     ));
   }
 
