@@ -3,13 +3,13 @@ import 'dart:async';
 import 'package:bloc/bloc.dart';
 import 'package:diet_driven/blocs/blocs.dart';
 import 'package:diet_driven/blocs/settings_edit/settings_edit.dart';
-import 'package:diet_driven/log_printer.dart';
+//import 'package:diet_driven/log_printer.dart';
 import 'package:diet_driven/repositories/repositories.dart';
-import 'package:logger/logger.dart';
+//import 'package:logger/logger.dart';
 import 'package:meta/meta.dart';
 
 class SettingsEditBloc extends Bloc<SettingsEditEvent, SettingsEditState> {
-  final logger = getLogger("settings edit bloc");
+//  final logger = getLogger("settings edit bloc");
   final String userId;
   final SettingsRepository settingsRepository;
 
@@ -25,7 +25,7 @@ class SettingsEditBloc extends Bloc<SettingsEditEvent, SettingsEditState> {
       try {
         await settingsRepository.updateDarkMode(userId, event.darkMode);
         event.completer?.complete();
-        logger.i("dark mode now ${event.darkMode}");
+//        logger.i("dark mode now ${event.darkMode}");
       } on Exception catch(e) {
         event.completer?.completeError(e);
       }
