@@ -1,8 +1,8 @@
 import 'dart:async';
 
 import 'package:bloc/bloc.dart';
-import 'package:diet_driven/log_printer.dart';
-import 'package:logger/logger.dart';
+//import 'package:diet_driven/log_printer.dart';
+//import 'package:logger/logger.dart';
 import 'package:meta/meta.dart';
 import 'package:rxdart/rxdart.dart';
 
@@ -17,7 +17,7 @@ import 'package:diet_driven/models/models.dart';
 /// Aggregates and manages authentication and settings.
 /// [UserDataBloc] shows loading or onboarding until loaded.
 class UserDataBloc extends Bloc<UserDataEvent, UserDataState> {
-  final logger = getLogger("user data bloc");
+//  final logger = getLogger("user data bloc");
   final UserRepository userRepository;
   final SettingsRepository settingsRepository;
 
@@ -28,7 +28,7 @@ class UserDataBloc extends Bloc<UserDataEvent, UserDataState> {
     assert(settingsRepository != null);
 
     final userDataEvent$ = userRepository.authStateChangedStream // .delay(Duration(seconds: 20))
-      .doOnData((user) => logger.d("USER: $user"))
+//      .doOnData((user) => logger.d("USER: $user"))
       // Side effect ensures user is authenticated and new user doesn't see userData from previous user
       .doOnData((user) {
         if (user == null) {
@@ -81,12 +81,12 @@ class UserDataBloc extends Bloc<UserDataEvent, UserDataState> {
         ..subscription = event.subscription
       );
 
-      logger.i("loaded user data ${event.authentication.uid}");
+//      logger.i("loaded user data ${event.authentication.uid}");
     }
     if (event is OnboardUser) {
       yield UserDataUnauthenticated();
 
-      logger.i("onboarding user");
+//      logger.i("onboarding user");
     }
     if (event is UserDataError) {
       yield UserDataFailed((b) => b
@@ -94,7 +94,7 @@ class UserDataBloc extends Bloc<UserDataEvent, UserDataState> {
         ..trace = event.trace
       );
 
-      logger.i("user data failed");
+//      logger.i("user data failed");
     }
   }
 }
